+++ src/max_div/_core/metrics/_distance/__init__.py
@@ -1,2 +1,9 @@
-from ._compute import compute_pdist, compute_separation, get_pdist_el, update_separation_add, update_separation_remove
+from ._compute import (
+    compute_pdist,
+    compute_separation,
+    get_pdist_el,
+    update_separation_add,
+    update_separation_remove,
+    validate_cosine_vectors,
+)
 from ._enum import DistanceMetric